Moves FormFieldAdvancedSelection
