+++ src/components/Field/index.tsx
@@ -4,3 +4,6 @@ export * from "./FieldTypes";
 
 export { default as FormFieldAddress } from "./FormFieldAddress";
 export * from "./FormFieldAddress";
+
+export { default as FormFieldAdvancedSelection } from "./FormFieldAdvancedSelection";
+export * from "./FormFieldAdvancedSelection";
