+++ src/components/Form/stories/Playground.stories.tsx
@@ -15,7 +15,7 @@ import Form from "../Form";
 
 // Types
 import { FieldDef } from "@root/components/Field";
-import { additionalOptions } from "@root/forms/FormFieldAdvancedSelection";
+import { additionalOptions } from "@root/components/Field/FormFieldAdvancedSelection";
 import { getOptionsCountries, getOptionsStates } from "@root/components/Field/FormFieldAddress/utils/optionGetters";
 import { nanoid } from "nanoid";
 import { columns, numberTableDefaultValue, rows } from "@root/forms/FormFieldNumberTable/numberTableUtils";
