+++ src/forms/FormFieldRadio/FormFieldRadio.test.tsx
@@ -6,7 +6,7 @@ import { getOptions } from "@root/utils/getOptions";
 // Components
 import FormFieldRadio from "./FormFieldRadio";
 import { MosaicLabelValue } from "@root/types";
-import { additionalOptions } from "../FormFieldAdvancedSelection";
+import { additionalOptions } from "../../components/Field/FormFieldAdvancedSelection";
 
 afterEach(cleanup);
